Make shill depend on a newer version of libchrome.
This is part of the upgrade of shill to use the newest
version of libbase.

Note that this CL does not actually have to go in at the
same time as https://gerrit.chromium.org/gerrit/16156,
because this ebuild change justs makes shill depend on
a new slotted version of libchrome, but it's the shill
Makefile that causes it to build against that version.

BUG=chromium-os:15330
TEST=Build shill, run it on device, run unit tests.

Change-Id: Id3ce4ab346ab2270e8e5d08b317fe9ec5618cd09
Reviewed-on: https://gerrit.chromium.org/gerrit/17883
Commit-Ready: Eric Shienbrood <ers@chromium.org>
Reviewed-by: Eric Shienbrood <ers@chromium.org>
Tested-by: Eric Shienbrood <ers@chromium.org>