openssl: drop -Wl,-rpath hardcodes
There is no point in telling binaries to use an rpath with the system
library since the ldso at runtime will search that.  This just causes
us problems when attempting to be tricky with LD_LIBRARY_PATH (as the
DT_RPATH tag will override that search path).

BUG=chromium-os:27683
TEST=`emerge-x86-alex openssl && readelf -d /build/x86-alex/usr/bin/openssl` shows no RPATH tags

Change-Id: I53f6129444e513f4d1d5bef96760c8dd79ae636f
Reviewed-on: https://gerrit.chromium.org/gerrit/17894
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>