Add CROS_WORKON_COMMIT to USE_EXPAND.
Currently, if the CROS_WORKON_COMMIT of an ebuild is changed, the
prebuilt is not invalidated. This causes prebuilts of any builder that
uprevs to be invalid. We should fix this so that any builder can uprev.

This also helps improve the correctness of trybots, by ensuring any
prebuilts stored by these builders locally are also valid.

Here's a situation today where this CL would help. Let's say you upload
a CL and try it on the remote trybot. It compiles fine but then fails
tests. You fix the CL and submit it to the commit queue. Commit queue
pushes your CL live. Later, I try a CL on the same remote trybot and
your package has the exact same PVR as the one on the commit queue,
but a different CROS_WORKON_COMMIT. In this situation, the remote
trybot would not update to a new version because it's not smart
enough to know that the CROS_WORKON_COMMIT differs between two ebuilds
the the same PVR.

Similarly, if full builders uprev, let's say you CHUMP in a commit to
one repository, and then send a second commit through the commit queue.
The full builder picks up the CHUMP commit and points at that
CROS_WORKON_COMMIT. The commit queue picks up both commits and submits
an ebuild that contains both. In that situation the full builder would
submit a prebuilt that has the same PVR as your original change. This
causes the builds to be subtly incorrect.

Both of the above situations are uncommon situations, but I was concerned
enough about them that I didn't want to enable uprevving on full builders
until the above was fixed. This CL fixes issues with CROS_WORKON_COMMIT
being out of sync.

BUG=chromium-os:27840
TEST=Test that prebuilts are invalidated if the CROS_WORKON_COMMIT
     changes.

Change-Id: I5742bfe2f4d2d5ff06cfd28e818a0367101a9b52
Reviewed-on: https://gerrit.chromium.org/gerrit/18223
Commit-Ready: David James <davidjames@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>