Reduce the Palm minimum distance on Cr48
We don't get different pressure readings for the two contacts, which is
a big part of thumb detection, here we have to reduce the palm minimum
value to fix the issue that two-finger scrolling not working with a resting
thumb on Cr48.

BUG=chromium-os:27850
TEST=device(on Cr48) and see if 2f scrolling with a resting thumb
is working.

Change-Id: If7d577eb5cb6623a2636004d7e12c53512690f20
Reviewed-on: https://gerrit.chromium.org/gerrit/18342
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Ready: Chung-yih Wang <cywang@chromium.org>
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Chung-yih Wang <cywang@chromium.org>