Sync Chrome in the distdir, but build in the cache dir.
This CL separates the distdir from the cache dir for Chrome. This allows
buildbots and users to use the same Chrome source across multiple chroots.
This should reduce the time required to run a full build.

CHROME_CACHE_DIR is used for storing output artifacts, and is always a
regular directory inside the chroot (i.e. it's never mounted in, so it's
always safe to use cp -al for these artifacts.)

CHROME_DISTDIR is used for storing the source code, if any source code
needs to be unpacked at build time (e.g. in the SERVER_SOURCE scenario.)
It will be mounted into the chroot, so it is never safe to use cp -al
for these files.

CHROME_ROOT is the location where the source code is used for compilation.
If we're in SERVER_SOURCE mode, CHROME_ROOT is CHROME_DISTDIR. In LOCAL_SOURCE
mode, this directory may be set manually to any directory. It may be mounted
into the chroot, so it is not safe to use cp -al for these files.

BUG=chromium-os:27988
TEST=Ran several remote trybot runs, and verified build is 5 minutes faster.

Change-Id: Ib38b09782a8251df54a33fd9c5936fb45141d996
Reviewed-on: https://gerrit.chromium.org/gerrit/18419
Commit-Ready: David James <davidjames@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>