cros-workon: Do not create the USE flag with default value
The goal of the USE flag is to force rebuilds of cros-workon packages on
trybots. However, simply the existence of the USE flag should be enough
to make --newuse rebuild.

On the other hand, having default values is not supported in EAPI=0
which breaks some packages.

BUG=none
TEST=build anything

Change-Id: I4291a9d3c99bc5648685d6f0dcd51208ae161f0f
Reviewed-on: https://gerrit.chromium.org/gerrit/18505
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Zdenek Behan <zbehan@chromium.org>
Commit-Ready: Zdenek Behan <zbehan@chromium.org>