Add to cros-binary.eclass functionality for local file testing.
This change formalizes a method for overriding the URL prefix
for binaries being fetched by cros-binary derivatives.
@ECLASS-VARIABLE: CROS_BINARY_LOCAL_URI_BASE can be defined in
a user's environment or in a locally-edited ebuild to facilitate
testing without the binary payload from its canonical (remote)
hosting.

The change was initially reviewed at
  https://gerrit.chromium.org/gerrit/19149
  Change Icbccb6f1: Add logging output to cros-binary.eclass for
      missing sha1 values

and was reverted due to a bad interaction with a private ebuild.
This changelist is identical save for an reworked commit message.

BUG=None
TEST=None

Change-Id: I1f5132797c69d02e1fa9b1a77df140b86876ee54
Reviewed-on: https://gerrit.chromium.org/gerrit/20120
Reviewed-by: Allen Martin <amartin@nvidia.com>
Commit-Ready: Micah Catlin <micahc@chromium.org>
Reviewed-by: Micah Catlin <micahc@chromium.org>
Tested-by: Micah Catlin <micahc@chromium.org>