gcc: optimize ccache usage a bit more
First, set a limit on the cache size.  Bots don't like it when you
take up 20GiB of space, and this exploded faster than I expected.

Next, improve the overall expiration of the cache dir.  Key off of
the gcc git hash to see when we need to nuke the whole dir or keep
it around.  This keeps the cache from expiring when it's not needed,
such as: install x86_64/i686 toolchains & do some building.  Later,
install the arm toolchain.  Currently, that'd clear the cache, but
now it'll clear it only based on the current target.  Further, if
we change the ebuild, but not the gcc source code, we don't need to
clear the cache as the gcc itself should produce the same code.

Finally, utilize CCACHE_BASEDIR so that we can share cached objects
across boards.  This doesn't preclude other checks (like flags),
just improves the hashing of file paths so the unique sysroot path
is not considered.

BUG=chromium-os:25613
TEST=`cbuildbot chromiumos-sdk` works

Change-Id: If8a4cc9cbf9443d742e99dab31b23b8a06ea07fc
Reviewed-on: https://gerrit.chromium.org/gerrit/19321
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>