Verify prebuilts using content hashes.
We should use the SHA1 of the contents of the source tree to validate
the prebuilts. Unlike the commit hash, this SHA1 is unaffected by the
history of the repository, or by commit messages. This ensures that
users always have correct prebuilts, without causing unnecessary
rebuilds.

Design document: http://goo.gl/yAio6

CQ-DEPEND=19426
BUG=chromium-os:28783
TEST=Remote trybot run.

Change-Id: I07d4ba2516756b4227e4885899955f82b1af095c
Reviewed-on: https://gerrit.chromium.org/gerrit/19427
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>