Disable redundant collision-protect flag.
parallel_emerge already disables collision-protect because it isn't
terribly useful and causes problems. I'd like to move this into our
configuration, since it causes issues regardless of parallel_emerge.

Portage has two flags for doing collision protection: collision-protect
and protect-owned. The protect-owned feature is enabled by default and
is quite useful: it checks to make sure that we don't have multiple
packages that own the same file. The collision-protect feature is more
strict, and less useful: it fails if it finds a conflicting file, even
if that file was created by an earlier ebuild that failed to install.

We want to disable collision-protect here because it doesn't deal with
emerge failing very well. If emerge crashes and leaves files around,
the collision-protect then fires and prevents you from ever installing
any packages again. This is bad, and not terribly useful.

BUG=chromium-os:28855
TEST=Trybot runs.

Change-Id: I8adbb10608f474c315308ccd83b416d2dcae7119
Reviewed-on: https://gerrit.chromium.org/gerrit/19503
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>
Commit-Ready: David James <davidjames@chromium.org>