bluez: retry bonding after autopairing failure
Add the second set of patches planned for R20 that permit a plugin to
add bonding callbacks and attempt a retry in the case of failure, e.g.
with the PIN Code that it generated.

Move the keyboard PIN generation into a plugin to take advantage of this;
now when we fail a bonding with a generated PIN, we retry the bonding and
fallback to prompting the user for the PIN instead.

BUG=chromium-os:28022
TEST=verified with existing devices and failing Motorola keyboard

Change-Id: I6eb30214640a3a3d0964062ad89110c73efca4e0
Reviewed-on: https://gerrit.chromium.org/gerrit/19924
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>