htpdate: allow to use 64-bit time delta
The previous code was artificially limited to a delta of 2100000000
seconds and would fail to set the time if the delta is larger.

When the hardware RTC registers are filled with random values due to
lack of power on the RTC/"always on" rail  (which is more frequent on systems
without CMOS backup battery), we need to be able to recover from a
situation where the current date is several hundred years in the future.

BUG=chrome-os-partner:8532
TEST=on Link proto-1 (without CMOS battery) check that ntpdate is able
to set the current date after we ran out of battery and filled RTC
registers with garbage.

Change-Id: Ia598e7d37ca28d7be15e42dfc780cbb5e00992ad
Reviewed-on: https://gerrit.chromium.org/gerrit/20342
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>