Added flush to stdout before execv of wrapped process
This is a known issue with Python handling of messages that go to stdout.
In Python stdout is buffered. When the output goes to a terminal, the stdout
is flushed on every newline. When the output sent to stdout is redirected
to a file it is undefined when stdout is flushed.
This results in undefined behaviour when sending the output of the sysroot
wrapper to a file.
Normally at the exit of the sysroot wrapper the stdout should be flushed.
However, sysroot wrapper does an execv. Python documentation clearly states
that before an execv one should do a flush of stdout in case there is any
pending data in there.
So, to fix the problem I am just adding the corresponding flush before execv.
I could have used the "-u" (unbuffered) Python option or other solutions
that do a flush after each "print" but I think adding the flush before
execv fixes the issue at the root of the problem.

BUG=chromium-os:28470
TEST=cbuildbot -g 'I22fdbac391de1af452682e1898de696d193f235b' chromiumos-sdk

Change-Id: I22fdbac391de1af452682e1898de696d193f235b
Reviewed-on: https://gerrit.chromium.org/gerrit/20350
Reviewed-by: asharif <asharif@chromium.org>
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>