Use kernel_noload on ARM fitimage kernels
Change the kernel image type to kernel_noload to avoid having to copy
the kernel one extra time on the target after loading.

This means that u-boot can load the fitImage anywhere, and it can run
from there. No need to hardcode a load_addr into it, that would otherwise
cause problems due to different ARM vendors having physical memory at
different locations. This way we can just load it anywhere and jump to
it, and the zImage decompressor takes care of the rest.

Since the u-boot we have on tegra platforms won't support the
kernel_noload version, let's keep that one at the old format. This is
a semi-temporary hack until we have moved everybody forward.

Also, add a current temporary lookup of board overlay to kernel devicetree
file name. This is temporary until we get the solution in u-boot where
we build and include _all_ supported device trees in the image and pick
the right one before launching the kernel.

BUG=chrome-os-partner:8942
TEST=boot an image on a daisy and kaen, make sure the fit image creation
output looks sane (modulo the dtc lack of support for kernel_noload
causing a warning).

Change-Id: Icc3583f8a763602cd511d92f23125b529f573877
Reviewed-on: https://gerrit.chromium.org/gerrit/20347
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Olof Johansson <olofj@chromium.org>