bluez: add PS3 Controller pairing plugin
The PS3 Bluetooth Controller must not only be paired using a USB cable
but also connects to the host itself when the PS button is pressed rather
than allowing connections from the host.

This necessitates a bluetoothd plugin to handle it, responding to the
USB plug event and creating the device in bluetoothd's database such that
future connections will be trusted; the device itself also needs
programming to contact the host when its PS button is pressed.

The plugin also handles the bluetooth connection, sending the appropriate
commands to update the lights on the controller.

BUG=chromium-os:30273
TEST=verified controller pairs when USB cable connected and connects via Bluetooth.

Change-Id: Ia75321235035d9a26d09c83b925ae32d743e0c9c
Reviewed-on: https://gerrit.chromium.org/gerrit/21668
Reviewed-by: <scottmg@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>