Remove redirection of sudoed unit test logs.
This redirection caused some confusion due to a failure among the subset
of privileged update engine unit tests. Without it, it would be possible
to actually diagnose what's causing failures on the buildbots.  As we
found no justification of why it might be necessary, it's being removed.

BUG=chromium-os:29841
TEST=Ebuild properly run unittests; sudo invocation log message appear
as expected.

Change-Id: If426ae1c1119c5650954c0fe38698abc357f14d0
Reviewed-on: https://gerrit.chromium.org/gerrit/21688
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Ready: Gilad Arnold <garnold@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>