Automate building netbootable firmware
Ensure that firmware legacy image includes networking capabilities for
platforms built with factory-mode and for Link - unconditionally.

To achieve that the u-boot needs to be built twice now, once without
networking enabled (for regular firmware) and once with networking
enabled (for factory imaging flow). The disadvantage of this method is
the need to build u-boot twice, this is why it is limited in scope. A
better solution might be building a netboot capable firmware only for
inclusion in the factory bundle, it would require more changes and
more thorough testing.

BUG=chrome-os-partner:7952
TEST=manual
 . run

   emerge-link chromeos-u-boot chromeos-bootimage

 . try booting /build/link/firmware/legacy_image.bin, observe that it
   starts netbooting while in u-boot
 . try booting /build/link/firmware/image.bin, observe that it
   properly starts up ChromeOS

Change-Id: Ied4336e44111e6639bf9b9b7a0c2a4ff975fe6d5
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/22075