Build legacy image using an alternative device tree if available
Managing serial console in the legacy image at compile time proved
poorly portable, it is much more robust when controlled through the
device tree. A new Link device tree has been added in a separate
u-boot CL, this change makes use of it.

If a dedicated device tree for the platform's legacy image exists -
use it instead of the standard device tree. If the dedicated device
tree does not exists - fall back to using the original one.

BUG=chrome-os-partner:7952
TEST=manual

   . with the appropriate u-boot and coreboot changes applied  run command
     emerge-link cheomeos-bootimage
   . boot a link system using /build/link/firmare/image.bin, observe
     it come all the way up to the ChromeOS login screen. Note that
     the serial port console keeps quiet and that u-boot and coreboot
     output is available in /sys/firmware/log
   . boot a link system using /build/link/firmare/legacy_image.bin and
     observe it come up trying to netboot with u-boot console active
   . run command 'emerge-stumpy chromeos-bootimage'. Observe the

  Dedicated device tree "/build/stumpy/firmware/dts/stumpy_legacy.dts" not found

     message in the emerge output

Change-Id: I2a5e13ad1fd4d984c05556f3cf8696b825ac57c4
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/22200
Reviewed-by: Vic Yang <victoryang@chromium.org>