chromeos: drop gtk module cache
The only things using gtk still are gtalk and o3d.  The former uses it for
a few tests, but not in the code itself.  The latter uses it for message
loops, not gui interaction, so they don't need the modules for input or
pixmaps.  That means we can drop the cache generation steps as they take
a while to run.

BUG=chromium-os:31095
TEST=build_packages+build_image for x86-alex works and boots

Change-Id: I785980364d1fe28a1560ba6c68584a070f91c733
Reviewed-on: https://gerrit.chromium.org/gerrit/23233
Reviewed-by: David James <davidjames@chromium.org>
Reviewed-by: Josh Horwich <jhorwich@google.com>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>