chromeos-fonts: move font cache generation here
We want to generate font caches after all the fonts we care about
have been merged.  This can be slow when using qemu, so move the
generation out of the main ebuild and to the one which depends on
all the font packages.  This way it can be run in parallel with
everything else, and we only regen caches when fonts change.

BUG=chromium-os:31094
TEST=build_packages+build_image for x86-alex boots and fonts render

Change-Id: I604a95d9567cf98f19af1bb84449abb7e0fb0946
Reviewed-on: https://gerrit.chromium.org/gerrit/23398
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>