x11-drivers/xf86-input-evdev: Fix crasher in scroll wheel accel patch
The amazing software Valgrind noticed that for some input devices, we
were walking past the end of an array allocated in the scroll wheel
accel code. Turns out we mandate that all mice get at least 4
valuators: WHEEL, HWHEEL START_TIME, END_TIME. The original patch
added 2 to the number of valuators we need, which works most of the
time b/c most mice would have gotten WHEEL and HWHEEL already. With
this CL, we add up to 4 to the number of valuators as necessary in
case the mouse originally would have gotten none of the ones we
require.

BUG=chromium-os:30712
TEST=tested on device w/ problematic mouse (MS Wireless Mobile Mouse
3500)

Change-Id: Id058e54f13167138c054f6034160fe8f8da85bb2
Reviewed-on: https://gerrit.chromium.org/gerrit/23455
Reviewed-by: <oscarpan@google.com>
Reviewed-by: Ang Pan <oscarpan@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>