power_manager : Reduce blank to lock delay on desktops
Reduce the black to lock delay as desktop systems will
have a longer time to resume external displays. If the
user is accustomed to entering their password right away
when the display is blanked, this will prevent cases
where the user accidentally types his password while the screen
is not locked.

10 seconds after the screen blanks, the screen will lock.

BUG=chromium-os:23913
TEST=Try on stumpy.
cat /usr/share/power_manager/react_ms
Check that it is 10000
cat /usr/share/power_manager/lock_ms
Check that it is 490000

Change-Id: Ib65cc534f00cec70793ff4952b423ffb427f0e90
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/23537
Reviewed-by: Simon Que <sque@chromium.org>