xorg-conf: mario-cmt: Enable Sensor Jump Filter
The mario profile sensor can cause significant "sensor jump" on the order
of ~14mm.  This is seen, for example, when a moving finger approaches a
stationary finger.  This causes many pointer jumps in logs, such as those
referenced by:
http://crosbug.com/31329.

Enabling the Sensor Jump Filter also has a good affect on drum roll,
when drumroll happens within the sensor jump range, such as the jumps in:
http://crosbug.com/31303

The "Sensor Jump Max Dist" was chosen by analying pulling affect.
It is the inter-finger distance at which the pulling affect becomes
significant.

The "Sensor Jump Filter" also has a positive affect on drumroll
suppression.  Drumroll jumps that look like rapid one-finger motion
that occur within this range will be also be surpressed.

BUG=chromium-os:31329,chromium-os:31303
TEST=(1) manual:
place resting thumb in bottom center of pad, moving finger back and forth
over it.  Should see no pointer jumps, and no teleporting.
(2) replay logs in chromium-os:31329 showing jumps and notice that the
jumps are gone.
(3) replay logs in chromium-os:31303 with drum-roll, and notice that the
jumps are gone.

Change-Id: I546e3e1cfa6c7b5ce93d19d4edcd1eedc0212588
Reviewed-on: https://gerrit.chromium.org/gerrit/23978
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>