Create an NSS db and put in /etc/skel
The first process to use libchrome_crypto's NssUtil class to access
the user's NSS DB (/home/chronos/user/.pki/nssdb) will wind up
creating the database before using it.  If this process is running as
a non-chronos user, this can result in Chrome (and other things) being
unable to generate keys and store certificates.  We haven't seen this
in the wild, but it is happening in some test scenarios.  In order to
defeat this race, we can pre-populate each user's cryptohome with an
empty NSS DB that's owned by the right user.

Placing an appropriate directory structure in /etc/skel will cause
cryptohomed to install the right stuff in the right place as it creates
the user's cryptohome.

BUG=chromium-os:30932
TEST=x86-alex remote trybot, build kaen image and run smoke suite on device.

Change-Id: Icac4fa3f6bda520ae8e1f2d02e9e548d33ab948a
Reviewed-on: https://gerrit.chromium.org/gerrit/24555
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>