app-misc/utouch-evemu: added resolution field for absinfo entries
evemu has been updated to write/read the resolution field of input_absinfo
structures. To preserve compability with old files the reading code
supports lines with and without the resolution field.
It also fixes a bug that turns out to be visible now when reading absinfo
entries: absinfo is never initialized to 0. So resolution will have
an undefined value if it is not read from the file.

The ebuild file has been updated to include the patch for this feature, as well as one to fix the bug.

BUG=chromium-os:30921
TEST=evemu-describe $DEV > test; evemu-play $DEV < test; Runs without errors.

Change-Id: I275bfbeec41bca4aed1e4e926d5385a305601fd6
Reviewed-on: https://gerrit.chromium.org/gerrit/24661
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Ready: Dennis Kempin <denniskempin@chromium.org>
Tested-by: Dennis Kempin <denniskempin@chromium.org>