chromeos-bootimage: Remove support for a separate legacy fdt file
This is not the correct way to enable a serial console. U-Boot now has
support for dealing with an existing fdt node, so we should be able to
punt this and simplify the logic before we make it more complicated
again.

BUG=chrome-os-partner:7969
TEST=manual
$ sudo rm /build/link/firmware/dts/link_legacy.dts
$ emerge-link chromeos-bootimage
See that it produces images correctly still

Change-Id: I0534148677ed3398291d0dd9c630a5c96aecbeda
Reviewed-on: https://gerrit.chromium.org/gerrit/25144
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>