chromeos-bootimage: Remove a few old comments
These comments are no longer relevant. We are hopefully not going to have
virtual ebuilds for chromeos-bootimage now that we have the new
cros_bundle_firmware. The bct files are certainly only used on Tegra,
but there is no harm in providing them to cros_bundle_firmware.

BUG=chromium-os:31643
TEST=manual:
$ emerge-link chromeos-bootimage

Change-Id: I1b2d5050c920722692f883e549dee34fb8b6b2f0
Reviewed-on: https://gerrit.chromium.org/gerrit/25150
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>