inputcontrol: Add X USE flag, chgrp input devices to chronos
The X flag is enabled by default. When it is disabled, apply a patch to
chgrp input devices to chronos, allowing Chrome to handle the input
processing instead.

BUG=chromium-os:32047
TEST=emerge inputcontrol with and without "USE=-X". Check these
installed files:
/build/$BOARD/opt/google/mouse/mouse_added
/build/$BOARD/opt/google/touchpad/touchpad_added
In the version without "USE=-X", they should be the same as in ToT.
In the version with "USE=-X", they should have an extra "chgrp chronos"
line.

Change-Id: I88df9fc1b0cc43608726af277a3a2dd7d83213e6
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26067