cros-kernel2.eclass: try fallback defconfigs if available
This change makes the kernel eclasses try to fall back to defconfigs saved
in the ebuild files directory for those kernel trees that lack (or fail to
configure) the regular kernel config files.

This is useful to make it trivial to test upstream kernels in our
build environment.

A later CL will move the current config fragments to be under the cros-kernel/
directory instead, since we now have a place to store them.

BUG=chromium-os:32217
TEST=remove chromeos in a kernel tree, try building for x86-mario, lumpy and daisy

Change-Id: I6c315617fa1bce0c612d2a11fdeb84438f1a04ae
Reviewed-on: https://gerrit.chromium.org/gerrit/26300
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: David James <davidjames@chromium.org>