Make chromeos-bootimage aware of butterfly/stout boards.
There's an enumeration of BOARD values here that has to be kept in sync
with the availability of new boards.  Since having board-specific logic
in ebuilds is exceptional, the long-term plan is to refactor these
ebuilds so that board-specific customizations live in the board specific
overlay, but for now carry on the current pattern of adding new boards
here as well as in cros-board.eclass.

BUG=chromium-os:32422
TEST=emerge-$BOARD chromeos-bootimage ; # for all $BOARD we build firmware for

Change-Id: Icb11dcf330e6e9300efe2d394ab4116e2ad6e9a9
Reviewed-on: https://gerrit.chromium.org/gerrit/26975
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Micah Catlin <micahc@chromium.org>
Tested-by: Micah Catlin <micahc@chromium.org>