cros-workon: support out-of-tree builds for non-9999 versions too
Since most people will have the code already checked out locally and at
the right revision, take advantage of that when the package has out of
tree support enabled.

We make sure the repo exists, is at the right rev, and contains no local
changes before we use it.  But this is the common state, so that's OK.
If the repo fails any of these tests, we just fall back to the normal
workflow so things continue to "just work".

BUG=chromium-os:30559
TEST=`emerge-x86-generic =sys-kernel/chromeos-kernel-3.4-r1167` uses existing checkout
TEST=`emerge-x86-generic sys-kernel/chromeos-kernel` with dirty tree used dedicated git clone
TEST=`cbuildbot amd64-generic-full` passed
TEST=`cbuildbot arm-generic-full` passed
TEST=`cbuildbot x86-generic-full` passed

Change-Id: If4b57151fababe24cdf78694b969a1db3361e606
Reviewed-on: https://gerrit.chromium.org/gerrit/28316
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>