In cros-coreboot.eclass, RDEPEND=!sys-boot/chromeos-coreboot
To enable incremental builds across the transition from
chromeos-coreboot.ebuild to
cros-coreboot.eclass+chromeos-coreboot-board.ebuild, use
a blocking RDEPEND value in the eclass to force the removal
of the old package.  This avoids a "file collision" error
in portage when emerging the new one.

BUG=chromium-os:32662
TEST=emerge-$board sys-boot/chromeos-coreboot && emerge-$board sys-boot/chromeos-coreboot-$board

Change-Id: I3138b73d2984f9e6bdaa37572de975b12be1456e
Reviewed-on: https://gerrit.chromium.org/gerrit/28520
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Micah Catlin <micahc@chromium.org>
Tested-by: Micah Catlin <micahc@chromium.org>