htpdate: ensure we retrieve all HTTP headers
We currently have 2 issues when the HTTP server is sending back large
headers :
first, we are truncating them at 1k (the current size of our headers)
even though the interesting ones might be beyond that limit.
In case our buffer is larger than the MTU, we need to ensure that we are
waiting for all the packets to arrive before parsing the headers.
So, set the headers size limit to 8kB and block the read on the socket
until we get all the packets.

BUG=chrome-os-partner:32975
TEST=from US and international IPs, execute "htpdate -d -w
www.google.com" and check we can properly read the "Date:" header even
when it is more than 2k bytes after the beginning of the response.

Change-Id: I645e4c327dbf2e265e7e3a5fdd0c7a34f43ee457
Reviewed-on: https://gerrit.chromium.org/gerrit/28703
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>