make sure DISPLAY is not set
A recent change in shill/libbase implicitly had the unittests connect
to an X server.  On dev machines, this works because $DISPLAY is set.
On the buildbots, this fails because there is no X for them to connect
to.  Make sure DISPLAY is not set on the dev machines either when they
run unittests.

BUG=chromium-os:27723
TEST=`cros_run_unit_tests --board x86-alex -p shill` failed to connect to X

Change-Id: I20f85c58e530e52b57500f4739f56f728f05ea24
Reviewed-on: https://gerrit.chromium.org/gerrit/28953
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>