Add Noto Sans fonts.
Noto is the name of font familes to cover all the Unicode characters.
In this CL, the first in series, Noto Sans and Noto Sans UI in 4 weights
(regular, bold, italic and bold italic) are added. More will come later.

Noto Sans and Noto Sans UI are identical at the mmeont and
they're in TTC to avoid the duplication. They may diverge down the road,
but they'll remain packaged in TTC to share as much data as possible.

Noto Sans UI will replace Chrome Droid Sans as the primary UI font on
ChromeOS.
See http://codereview.chromium.org/10836082 for the Chromium-side
change.

1. Add a new package 'notofonts' under media-fonts. (FONT_SUFFIX set to
ttc)
2. Make chromeos-fonts depend on notofonts.
3. Add 'X' use mask for notofonts package

BUG=chromium:126735
TEST=./build-package notofonts works fine and the final build has both Noto Sans and Noto Sans UI available in Chrome's font selection menu.

Change-Id: I6ecb29fdffc8681dbfef91e88405c485390a03ad
Reviewed-on: https://gerrit.chromium.org/gerrit/29087
Reviewed-by: Jungshik Shin (jungshik at google) <jshin@chromium.org>
Tested-by: Jungshik Shin (jungshik at google) <jshin@chromium.org>