Check stay_awake_with_headphones flag.
This new use flag is used to change the suspend behavior if headphones are
attached.  It will be set for Stumpy so that the loud buzz when suspended is
avoided.

Make power_manager depend on adhd so it can use the cras library to ask
if headphones are plugged in.  Also a future commit will use libcras to
check if audio is playing so we don't suspend when using a USB device.

BUG=chrome-os-partner:10413
TEST=Leave stumpy idle while headphones attached, make sure it doesn't
suspend.  Do the same for lumpy, checking that it still suspends.  Also
check that Lumpy's lid still suspends.

Change-Id: I619bbd1c05b93f29ecb755f52c468e2f73476944
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29329
Reviewed-by: Benson Leung <bleung@chromium.org>