cros-workon: support building common.mk packages out-of-tree
The common.mk build system supports building out-of-tree, so start a
framework to build all these packages that way.  Gives a minor speed
up as we no longer have to do a git clone+checkout in the common case.

This also centralizes the common.mk plumbing so that packages based on
it don't have to specify the same stuff in every ebuild.

This doesn't directly impact packages as the functions aren't exported
yet.  Once we convert all the existing ebuilds over in follow up CLs,
we can export these so they become the default.

BUG=chromium-os:33327
TEST=`cbuildbot {x86,amd64,arm}-generic-full` worked

Change-Id: I2d10d10e6b46cc6e4f398898829c5e53cf0aab94
Reviewed-on: https://gerrit.chromium.org/gerrit/29482
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>