mesa: glsl: Restrict opt_dead_code_local to variable assignment
This optimization step behaves incorrectly when performing
assignments into array values which are later used as vectors.
This odd configuration is produced by Chrome as a result of its
GLSL translation layer when single and multiple elements of
vectors are referenced.

As a temporary fix to this problem, restrict the optimization to
assignments that write directly into a variable.

BUG=chromium-os:32326
TEST=Manually run Airmech; see sign in screen rather than grey box.

Change-Id: Ia62775b59c1ab4f408c43b5f658a8f01fd9b2bb5
Reviewed-on: https://gerrit.chromium.org/gerrit/29531
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: Chris Wolfe <cwolfe@chromium.org>
Tested-by: Chris Wolfe <cwolfe@chromium.org>