Fix plumbing of silent_console in the chromeos-bootimage ebuild.
The variable being used to compute the silent_console flags being passed to
cros_bundle_firmware was not the same as the one which was actually passed to
other parts of the script. silent_console was always dropped on the floor as a
result. This change fixes that problem.

BUG=chrome-os-partner:12074
TEST=Built with the cros-debug use flag disabled and saw that the console was
silenced.

Change-Id: I704217837e156e1ded19d64888dc3e23ce529a1a
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/29683
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>