ChromeOS: Add dev-util/quipper to the base image.
This is needed for profiling on live ChromeOS machines. Note that this does not
run quipper on the base machines. We need to go through the proper privacy team
approval before making such a change. This makes it easy for a developer to run
quipper manually to send perf data to us.

BUG=none
TEST=cbuildbot -g <CL#> --remote chromiumos-sdk

Change-Id: I067730abe1717bafbea135dd8ef2f9cc9de03b12
Reviewed-on: https://gerrit.chromium.org/gerrit/29719
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Matt Redmond <mrdmnd@chromium.org>
Commit-Ready: asharif <asharif@chromium.org>
Tested-by: asharif <asharif@chromium.org>