This fixes NSS so that MITM-style root certs work across logins.
NSS has two copies of the MITM root cert: one copy is a temporary
cert, which is created before the real NSS databases is loaded and is
not trusted, and the other copy is a permanent cert and is trusted.
While verifying certs, NSS incorrectly converts the trusted permanent
cert to the untrusted temporary cert.  This fixes the NSS bug.

BUG=chromium:141768
TEST=ran on device, tested with MITM proxy

Change-Id: Ic975063ea77a422036417243ad2883349a7027d2
Reviewed-on: https://gerrit.chromium.org/gerrit/31295
Reviewed-by: Wan-Teh Chang <wtc@chromium.org>
Reviewed-by: Zelidrag Hornung <zelidrag@chromium.org>
Commit-Ready: Greg Spencer <gspencer@chromium.org>
Tested-by: Greg Spencer <gspencer@chromium.org>