CHROMIUM: openssl: Disable TLS heartbeat by default
Add a use flag for the OpenSSL TLS heartbeat extension,
defaulting to false.  This will disable the heartbeat
extension, which seems to have interoperability issues
with some RADIUS servers.

BUG=chromium-os:33832
TEST=Run WiFiSecMat/072CheckWPA_1x_PEAP.  Compare a client
side tcpdump capture of the Client Hello packet with that
with this change, and note that the "Extension: Unknown 15"
extension header is no longer there.

Change-Id: I90e59cd0ddb17e89eab1ffdae72c98ef6767f99f
Reviewed-on: https://gerrit.chromium.org/gerrit/31316
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>