Update debug log to include absinfo & slot information
The absinfo structures tell us the range and resolution of each axis.
Also, when interpreting the debug log we won't know which slot is
active until we see a ABS_MT_SLOT event. We could discard data until
the first ABS_MT_SLOT event but this may result in loss of most of the
log. This adds the current slot to each event to avoid this problem.

BUG=chrome-os-partner:13887
TEST=run feedback reports, check feedback logs

Change-Id: I9ab25fe2afaef3c56c2de29ea48d7f3d21765c31
Reviewed-on: https://gerrit.chromium.org/gerrit/32821
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Tested-by: Michael Spang <spang@chromium.org>
Commit-Ready: Michael Spang <spang@chromium.org>